backend/requirements.txt
aiortc==1.10.1
av==13.1.0
# df==0.0.1
librosa==0.10.2.post1
nltk==3.9.1
numpy==2.2.4
openai==1.68.2
python-dotenv==1.1.0
rich==13.9.4
sounddevice==0.5.1
soundfile==0.13.1
soxr==0.5.0.post1
torch==2.6.0
torchaudio==2.6.0
transformers==4.49.0
websockets
